+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## 0.2.2
+* `SnapshotReadService` now allows to set a local snapshot file to read from instead retrieving it from AWS S3.   
+This functionality was moved from `SnapshotEventSource` and also works for a `CompactingKinesisEventSource` now.
+
 ## 0.2.1
 * Provide ```EncryptedOrPlainJsonTextEncryptor``` that checks whether data is plain json or is encrypted.